consolidate cooling solver into quokka::cooling
