+++ src/Cooling/test_cooling.cpp
@@ -51,6 +51,11 @@ template <> struct Physics_Traits<CoolingTest> {
 	static constexpr bool is_mhd_enabled = false;
 };
 
+template <> struct SimulationData<CoolingTest> {
+	quokka::cooling::cloudy_tables cloudyTables;
+	std::unique_ptr<amrex::TableData<Real, 3>> table_data;
+};
+
 constexpr double Tgas0 = 6000.;	      // K
 constexpr amrex::Real T_floor = 10.0; // K
 constexpr double rho0 = 0.6 * m_H;    // g cm^-3
@@ -59,11 +64,6 @@ constexpr double rho0 = 0.6 * m_H;    // g cm^-3
 const int kmin = 0;
 const int kmax = 16;
 Real const A = 0.05 / kmax;
-
-template <> struct SimulationData<CoolingTest> {
-	cloudy_tables cloudyTables;
-	std::unique_ptr<amrex::TableData<Real, 3>> table_data;
-};
 
 template <> void RadhydroSimulation<CoolingTest>::preCalculateInitialConditions()
 {
@@ -189,72 +189,10 @@ AMRSimulation<CoolingTest>::setCustomBoundaryConditions(const amrex::IntVect &iv
 	}
 }
 
-struct ODEUserData {
-	amrex::Real rho;
-	cloudyGpuConstTables tables;
-};
-
-AMREX_GPU_HOST_DEVICE AMREX_FORCE_INLINE auto user_rhs(Real /*t*/, quokka::valarray<Real, 1> &y_data, quokka::valarray<Real, 1> &y_rhs, void *user_data) -> int
-{
-	// unpack user_data
-	auto *udata = static_cast<ODEUserData *>(user_data);
-	Real rho = udata->rho;
-	cloudyGpuConstTables &tables = udata->tables;
-
-	// compute temperature (implicit solve, depends on composition)
-	Real Eint = y_data[0];
-	Real T = ComputeTgasFromEgas(rho, Eint, quokka::EOS_Traits<CoolingTest>::gamma, tables);
-
-	// compute cooling function
-	y_rhs[0] = cloudy_cooling_function(rho, T, tables);
-	return 0;
-}
-
-void computeCooling(amrex::MultiFab &mf, const Real dt_in, cloudy_tables &cloudyTables)
-{
-	BL_PROFILE("RadhydroSimulation::computeCooling()")
-
-	const Real dt = dt_in;
-	const Real reltol_floor = 0.01;
-	const Real rtol = 1.0e-4; // not recommended to change this
-
-	auto tables = cloudyTables.const_tables();
-
-	// loop over all cells in MultiFab mf
-	for (amrex::MFIter iter(mf); iter.isValid(); ++iter) {
-		const amrex::Box &indexRange = iter.validbox();
-		auto const &state = mf.array(iter);
-
-		amrex::ParallelFor(indexRange, [=] AMREX_GPU_DEVICE(int i, int j, int k) noexcept {
-			const Real rho = state(i, j, k, HydroSystem<CoolingTest>::density_index);
-			const Real x1Mom = state(i, j, k, HydroSystem<CoolingTest>::x1Momentum_index);
-			const Real x2Mom = state(i, j, k, HydroSystem<CoolingTest>::x2Momentum_index);
-			const Real x3Mom = state(i, j, k, HydroSystem<CoolingTest>::x3Momentum_index);
-			const Real Egas = state(i, j, k, HydroSystem<CoolingTest>::energy_index);
-
-			Real Eint = RadSystem<CoolingTest>::ComputeEintFromEgas(rho, x1Mom, x2Mom, x3Mom, Egas);
-
-			ODEUserData user_data{rho, tables};
-			quokka::valarray<Real, 1> y = {Eint};
-			quokka::valarray<Real, 1> abstol = {reltol_floor * ComputeEgasFromTgas(rho, T_floor, quokka::EOS_Traits<CoolingTest>::gamma, tables)};
-
-			// do integration with RK2 (Heun's method)
-			int steps_taken = 0;
-			rk_adaptive_integrate(user_rhs, 0, y, dt, &user_data, rtol, abstol, steps_taken);
-
-			const Real Eint_new = y[0];
-			const Real dEint = Eint_new - Eint;
-
-			state(i, j, k, HydroSystem<CoolingTest>::energy_index) += dEint;
-			state(i, j, k, HydroSystem<CoolingTest>::internalEnergy_index) += dEint;
-		});
-	}
-}
-
-template <> void RadhydroSimulation<CoolingTest>::computeAfterLevelAdvance(int lev, amrex::Real /*time*/, amrex::Real dt_lev, int /*ncycle*/)
+template <> void RadhydroSimulation<CoolingTest>::addStrangSplitSources(amrex::MultiFab &state, int /*lev*/, amrex::Real /*time*/, amrex::Real dt_lev)
 {
 	// compute operator split physics
-	computeCooling(state_new_cc_[lev], dt_lev, userData_.cloudyTables);
+	quokka::cooling::computeCooling<CoolingTest>(state, dt_lev, userData_.cloudyTables, T_floor);
 }
 
 auto problem_main() -> int
@@ -292,7 +230,7 @@ auto problem_main() -> int
 	sim.checkpointInterval_ = -1;
 
 	// Read Cloudy tables
-	readCloudyData(sim.userData_.cloudyTables);
+	quokka::cooling::readCloudyData(sim.userData_.cloudyTables);
 
 	// Set initial conditions
 	sim.setInitialConditions();
